services/indexer/pyproject.toml
[project]
name = "indexer"
version = "0.1.0"
description = "FineWeb indexer service with parallel workers"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "azure-storage-blob",
    "azure-servicebus",
    "fastapi",
    "uvicorn",
    "pydantic",
    "python-dotenv",
    "requests"
]
